+++ Runtime/Server/Analytics/ServerAnalyticsService.cs
@@ -5,8 +5,8 @@
 using AccelByte.Api;
 using AccelByte.Core;
 using AccelByte.Models;
-using NUnit.Framework;
 using System.Collections.Generic;
+using UnityEngine.Assertions;
 
 namespace AccelByte.Server
 {